Fix error with maps page rendering, caused by #121
