+++ src/pages/maps/@id/+onBeforeRender.ts
@@ -1,0 +1,28 @@
+import { PageContextBuiltInServer } from "vike/types";
+import { apiV2Prefix } from "~/settings";
+
+const apiAddress = apiV2Prefix + "/defs/sources";
+
+export async function onBeforeRender(pageContext: PageContextBuiltInServer) {
+  const { id } = pageContext.routeParams;
+
+  const params = new URLSearchParams({
+    format: "geojson",
+    source_id: id,
+  });
+  const response = await fetch(apiAddress + "?" + params);
+  const data: any = await response.json();
+  const map = data?.success?.data?.features[0];
+
+  return {
+    pageContext: {
+      pageProps: {
+        map,
+      },
+      documentProps: {
+        // The page's <title>
+        title: map.properties.name,
+      },
+    },
+  };
+}
